backend/src/synthetic_example_generation/_archive/machine-labeler/pyproject.toml
[tool.poetry]
name = "machine_labeler"
version = "0.1.0"
description = "Machine Labeler"
authors = ["IBM Research"]
maintainers = ["Erik Miehling <erik.miehling@ibm.com>"]
readme = "README.md"
license = "Apache-2.0"
packages = [
    { include = "machine_labeler", from = "src" }
]

# basic install
[tool.poetry.dependencies]
python = "^3.10"
langchain = "*"
numpy = "*"
openai = "*"
python-dotenv = "*"
pyyaml = "*"
tqdm = "*"
transformers = "^4.47.0"
tokenizers = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"